requirements.txt
Django>=4.2,<5.0
django-ckeditor-5>=0.2.10
Pillow>=10.0
python-dotenv>=1.0
whitenoise>=6.0
